pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "worldengine"
version = "0.1.0"
description = "Epic worldbuilding as code — CI-driven validation for fictional worlds"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "instructor>=1.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
worldengine = "engine.cli:main"
